+++ src/sonovault/__init__.py
@@ -3,6 +3,7 @@
 from ._version import __version__
 from .client import SonoVault
 from .errors import SonoVaultError
+from .pagination import paginate
 from .webhooks import verify_webhook_signature
 
-__all__ = ["SonoVault", "SonoVaultError", "verify_webhook_signature", "__version__"]
+__all__ = ["SonoVault", "SonoVaultError", "paginate", "verify_webhook_signature", "__version__"]